fix: address second reviewer's findings (codex)
A second reviewer posted four inline comments that I missed on the first pass —
my comment fetch was truncated and only surfaced CodeRabbit's. All four hold.

- Two unit suites could not run without database credentials. The round-trip test
  imported SURVEYJS_TEMPLATES from DatabaseSeedingUtils, which transitively builds
  a Supabase admin client at module evaluation; and the gated due-date test's
  static import of TestingUtils ran before describe.skip could gate it, so an
  optional integration file FAILED instead of skipping. Both passed locally only
  because next/jest loads .env.local. Templates moved to a side-effect-free
  scripts/surveyTemplates.ts; the integration helpers now load lazily inside the
  gate. Verified with .env.local moved aside: round-trip passes, integration skips.

- Clearing a numeric choice lost its type. nextChoiceValue inferred the type from
  the transient input value, so select-all/delete/retype turned a Likert 3 into
  "3" and orphaned every response stored against it — the exact failure this PR
  exists to prevent. Booleans were never preserved at all. The type is now
  remembered on the Choice, independent of the transient value, and never
  serialized. Extracted to factories so it is testable; 7 tests cover it.

- Karma is not transferred when staff toggle a thread's anonymity. Per-identity
  karma leaves existing likes credited to the old profile while the byline reads
  the new one, and a later unlike debits the new profile, hits the GREATEST(0,…)
  floor, and leaves the old counter permanently inflated. Reproduced end to end;
  A-B against the old normalized trigger confirms this is new with our change
  rather than pre-existing. Fixed with an AFTER UPDATE OF author trigger guarded
  by WHEN (OLD.author IS DISTINCT FROM NEW.author), chosen over patching the RPC
  so the invariant holds for any writer of the column.

Found while verifying and NOT fixed here: toggle_discussion_thread_author_anonymity
declares its author variables as text and compares them to uuid columns, so it
raises "operator does not exist: uuid = text" and has never worked, despite a live
UI caller. Pre-existing and outside karma, so it is reported rather than bundled.

Also fixes an unused import that broke the production build after extracting
nextChoiceValue — caught because I checked the build's exit code rather than
grepping its log, which had misled me once.

Verified: jest 1097 passing with the same two pre-existing failures; tsc at the
pre-existing 8; lint and prettier clean; db reset from scratch; surveys + karma +
audit/leaderboard 76 passed and office-hours 10 passed across chromium and webkit;
gated due-date integration tests green. The karma transfer was A-B checked by
dropping the trigger and confirming the new test fails on the transfer assertion.
